# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: BNI Finance API
  version: "1.0.0"
  description: |
    REST API untuk BNI — invoice keanggotaan
    (pendaftaran & renewal), pembayaran, member, chapter, dan pelaporan.

    ## Autentikasi

    Semua endpoint `/api/v1/**` butuh token, **kecuali** yang ditandai
    `security: []` — login, halaman bayar publik, dan webhook.

    ```
    Authorization: Bearer <token>
    ```

    Token didapat dari `POST /api/v1/auth/login` dan berlaku 12 jam (bisa
    diubah lewat `TOKEN_TTL`). Tidak ada sesi di server: token yang sudah
    diterbitkan berlaku sampai kedaluwarsa, jadi keluar (logout) cukup dengan
    membuang token di sisi klien.

    ## Peran

    | Peran | Boleh |
    |---|---|
    | `admin` | Semuanya |
    | `user` | Hanya **pembacaan** (`GET`) |

    Setiap penulisan oleh peran `user` dibalas **403**. Ini batas yang
    sebenarnya — cek peran di UI hanya menyembunyikan tombol.

    ## Bentuk respons

    Daftar selalu memakai amplop yang sama:

    ```json
    { "data": [ … ], "meta": { "total": 47, "limit": 50, "offset": 0 } }
    ```

    `data` selalu larik — tidak pernah `null` — meski kosong.
    Galat selalu `{ "error": "pesan dalam Bahasa Indonesia" }`.

    Seluruh field JSON memakai **camelCase**, sama persis dengan tipe di
    `src/types`, sehingga bisa dipakai klien TypeScript tanpa lapisan pemetaan.

  license:
    name: Internal — BNI
servers:
  - url: http://localhost:8080
    description: Pengembangan lokal
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Tukar kode OTP dengan token
      description: |
        Langkah kedua login, dipakai bila `/auth/login` menjawab
        `otpRequired: true`.

        **Kode enam digit hanya sejuta kemungkinan.** Yang membuatnya aman bukan
        panjangnya melainkan tiga batas di sekelilingnya: umurnya 10 menit,
        sekali pakai, dan percobaannya dibatasi lima per kode. Tanpa batas
        percobaan, menebaknya selesai dalam hitungan detik.

        Hitungan percobaan disimpan **di basis data**, bukan di memori proses —
        hitungan yang hilang tiap restart bisa dikosongkan penyerang dengan
        memaksa server restart. Pemeriksaan dan pencatatannya terjadi dalam satu
        transaksi dengan `FOR UPDATE`; terpisah, sepuluh tebakan yang dikirim
        bersamaan semuanya membaca hitungan yang sama dan batas lima tidak
        pernah berlaku.

        Sisa percobaan **disebutkan** dalam pesan galat. Tanpa itu orang
        mengetik ulang sampai terkunci tanpa peringatan, lalu meminta kode baru
        berulang kali — yang justru menghasilkan lebih banyak email dan lebih
        banyak kode hidup sekaligus.

        Email tak dikenal dan kode salah dijawab sama.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        "200":
          description: Kode benar; token diterbitkan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "400": { description: Kode salah, kedaluwarsa, atau jatah percobaannya habis }
        "429": { description: Terlalu banyak percobaan untuk akun ini }
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Minta tautan atur ulang kata sandi
      description: |
        **Selalu menjawab 200 dengan pesan yang sama**, terdaftar atau tidak.

        Membedakannya akan mengubah endpoint ini menjadi alat pemeriksa
        keanggotaan: siapa pun bisa mencoba daftar alamat dan mengetahui mana
        yang punya akun di sini.

        Dibatasi per email dengan pembatas yang sama seperti login. Tanpa itu,
        endpoint ini adalah tombol kirim-email gratis — bisa dipakai membanjiri
        kotak masuk orang lain, menghabiskan kuota kirim, sampai alamat
        pengirimnya ditandai spam. Kerusakannya menimpa orang yang tidak
        melakukan apa-apa.

        Tautannya berlaku 30 menit dan hanya bisa dipakai sekali. Yang disimpan
        di basis data adalah HASH tokennya, bukan tokennya — satu dump basis
        data karena itu tidak menghasilkan satu pun tautan yang bisa dipakai.

        Menjawab **503** bila pengiriman email belum dikonfigurasi.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Permintaan diterima (tidak menyatakan emailnya terdaftar)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "429": { description: Terlalu banyak permintaan untuk email ini }
        "503": { description: Pengiriman email belum dikonfigurasi }
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Tukar tautan reset dengan kata sandi baru
      description: |
        Token kedaluwarsa, token yang sudah dipakai, dan token yang tidak pernah
        ada dijawab dengan pesan yang **sama persis**. Membedakannya memberi
        tahu penyerang bahwa sebuah token pernah sah.

        Penandaan token dan penulisan kata sandi terjadi dalam satu transaksi.
        Terpisah, kegagalan di antaranya meninggalkan salah satu dari dua
        keadaan yang sama-sama buruk: token terpakai tapi kata sandi belum
        berubah (pengguna terkunci, tautannya sudah mati), atau kata sandi
        berubah tapi tokennya masih hidup.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 6 }
      responses:
        "200":
          description: Kata sandi berhasil diubah
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "400": { description: Token tidak berlaku, atau kata sandi terlalu pendek }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Masuk dan dapatkan token
      description: |
        Email dibandingkan tanpa membedakan huruf besar/kecil.

        Email tak dikenal dan kata sandi salah menghasilkan **pesan yang sama
        persis** dan waktu pemrosesan yang setara — akun tidak bisa dienumerasi
        lewat endpoint ini.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
            example: { email: "admin@bnifinance.local", password: "rahasia" }
      responses:
        "200":
          description: Berhasil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Email atau kata sandi salah
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example: { error: "email atau kata sandi salah" }
  /api/v1/auth/quick-login:
    get:
      tags: [Auth]
      summary: Daftar akun masuk cepat
      description: |
        Akun yang boleh masuk **tanpa kata sandi**, untuk demo dan pengembangan
        lokal. Hanya nama, email, dan peran — tidak pernah kredensial.

        Diatur lewat `AUTH_QUICK_LOGIN`, berisi daftar email dipisah koma dan
        **kosong secara bawaan**. Sengaja daftar email, bukan saklar on/off:
        saklar berarti sekali dinyalakan di produksi, *semua* akun jadi bisa
        dimasuki tanpa kata sandi. Di sini akun yang boleh harus disebut satu
        per satu.

        Menjawab **404** saat fitur mati — produksi tidak memberi petunjuk
        bahwa rute ini ada.
      security: []
      responses:
        "200":
          description: Daftar akun yang diizinkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/AuthUser" }
        "404":
          description: Quick login tidak aktif
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Auth]
      summary: Masuk cepat tanpa kata sandi
      description: |
        Menerbitkan token biasa untuk akun yang ada di daftar izin. Tokennya
        sama persis dengan hasil `POST /auth/login` — tidak ada sesi kelas dua.

        **404** bila fitur mati, **403** bila email di luar daftar izin. Server
        mencatat peringatan saat start selama fitur ini menyala.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { title: Email, type: string, format: email }
            example: { email: "admin@bnifinance.local" }
      responses:
        "200":
          description: Berhasil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "403":
          description: Akun tidak terdaftar untuk quick login
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Quick login tidak aktif, atau akunnya belum ada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Profil pengguna yang sedang masuk
      description: |
        Dibaca ulang dari database, bukan dari isi token — perubahan peran
        terlihat pada request berikutnya, bukan pada login berikutnya.
      responses:
        "200":
          description: Profil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Auth]
      summary: Ubah nama sendiri
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { title: Nama, type: string, minLength: 1, example: "Budi Santoso" }
      responses:
        "200":
          description: Profil setelah diubah
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/password:
    put:
      tags: [Auth]
      summary: Ubah kata sandi sendiri
      description: |
        `currentPassword` **wajib**. Tanpa itu, token yang dicuri sudah cukup
        untuk mengambil alih akun secara permanen.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePasswordInput" }
      responses:
        "204": { description: Berhasil, tanpa isi }
        "400":
          description: Kata sandi baru kurang dari 6 karakter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Kata sandi saat ini salah
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/users:
    get:
      tags: [Auth]
      summary: Daftar pengguna
      description: "**Admin saja.**"
      responses:
        "200":
          description: Daftar pengguna
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/AuthUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Auth]
      summary: Tambah pengguna
      description: "**Admin saja.** Peran default `user` bila tidak dikirim."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserInput" }
      responses:
        "201":
          description: Pengguna dibuat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Email sudah terdaftar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/users/{id}/role:
    patch:
      tags: [Auth]
      summary: Ubah peran pengguna
      description: |
        **Admin saja.** Menurunkan admin **terakhir** ditolak dengan 409 —
        itu keadaan yang tidak bisa dipulihkan lewat API.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { title: Peran, $ref: "#/components/schemas/UserRole" }
      responses:
        "200":
          description: Pengguna setelah diubah
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Ini satu-satunya admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/users/{id}/password:
    put:
      tags: [Auth]
      summary: Reset kata sandi pengguna lain
      description: |
        **Admin saja.** Tidak perlu kata sandi lama, tapi **tidak bisa dipakai
        pada diri sendiri** — untuk itu gunakan `PUT /auth/password`.
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword: { title: Kata Sandi Baru, type: string, minLength: 6 }
      responses:
        "204": { description: Berhasil, tanpa isi }
        "400":
          description: Kata sandi terlalu pendek, atau id adalah diri sendiri
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/users/{id}:
    delete:
      tags: [Auth]
      summary: Hapus pengguna
      description: |
        **Admin saja.** Menghapus diri sendiri atau admin terakhir ditolak
        dengan 409.
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: Terhapus }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Akun sendiri, atau admin terakhir
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ===========================================================================
  # Invoice
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token dari `POST /api/v1/auth/login`.
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Input tidak valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Token tidak disertakan, tidak valid, atau sudah kedaluwarsa
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            tanpaToken: { value: { error: "token tidak disertakan" } }
            kedaluwarsa: { value: { error: "sesi sudah berakhir — masuk kembali" } }
    Forbidden:
      description: Butuh peran admin
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "butuh akses admin" }
    NotFound:
      description: Data tidak ditemukan
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "data tidak ditemukan" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Pesan dalam Bahasa Indonesia. }
    ListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          description: Selalu larik — tidak pernah `null`, meski kosong.
          items: {}
        meta:
          type: object
          required: [total, limit, offset]
          properties:
            total:
              type: integer
              description: Jumlah baris yang cocok dengan filter, bukan hanya halaman ini.
            limit: { type: integer }
            offset: { type: integer }

    # --- Auth ----------------------------------------------------------------
    UserRole:
      title: Peran
      type: string
      enum: [admin, user]
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id: { title: ID, type: string, format: uuid }
        name: { title: Nama, type: string }
        email: { title: Email, type: string, format: email }
        role: { title: Peran, $ref: "#/components/schemas/UserRole" }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { title: Email, type: string, format: email }
        password: { title: Kata Sandi, type: string }
    LoginResult:
      type: object
      required: [user]
      properties:
        token:
          type: string
          description: >
            JWT HS256, dikirim sebagai `Authorization: Bearer <token>`.


            **Kosong bila `otpRequired` true** — kata sandi yang benar saja
            belum menerbitkan token saat OTP menyala.
        expiresAt: { type: string, format: date-time }
        otpRequired:
          type: boolean
          description: >
            `true` berarti kata sandinya benar tapi belum cukup: kode sudah
            dikirim ke email, dan token baru terbit setelah kode itu
            diverifikasi lewat `/auth/verify-otp`.


            OTP hanya berlaku bila sakelar `login_otp_enabled` menyala DAN SMTP
            terkonfigurasi. Dua syarat, bukan satu: sakelar menyala sementara
            email mati berarti tidak seorang pun bisa masuk — termasuk admin
            yang harus mematikan sakelarnya.
        user: { $ref: "#/components/schemas/AuthUser" }
    CreateUserInput:
      type: object
      required: [email, password, name]
      properties:
        email: { title: Email, type: string, format: email }
        password: { title: Kata Sandi, type: string, minLength: 6 }
        name: { title: Nama, type: string }
        role:
          title: Peran
          allOf: [{ $ref: "#/components/schemas/UserRole" }]
          default: user
    UpdatePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { title: Kata Sandi Saat Ini, type: string }
        newPassword: { title: Kata Sandi Baru, type: string, minLength: 6 }

    # --- Invoice -------------------------------------------------------------
...